Sets up a basic Hello World type program in pyscript
Since setting pyscript by itself takes a bit of time, this commit acts as basic Hello World, so everyone can start coding directly.
The __init__.py is a flask server.
The files in static would be our frontend assets.
The files in templates would be our html, since flask demands it in such a way for it's render_template function to work